finaltest/requirements.txt
whispercpp
openai
pynput
playsound
pvrecorder
numpy
git+https://github.com/aarnphm/whispercpp.git
gTTS
opencv-python
hume
hume[stream]
colorama
psycopg2
python-dotenv